More accurate CD/DVD emulation in USB gadget
This fixes a number of things. Firstly, it adds support for DVDs,
so we no longer have the ~2GB limit on disk sizes. However, there
is also a significant improvement in the completness of the ATAPI
commands. Specifically, we now implement 'READ TOC (format=0x01)',
'READ DISK INFORMATION', and much of 'GET CONFIGURATION'. Currently
we announce support for the features required of the DVD profile
in the 'GET CONFIGURATION' response.